demo_app/backend/requirements.txt
fastapi
uvicorn
httpx
PyNaCl
base58
